+++ src/Symfony/Component/Messenger/Tests/Transport/AmqpExt/AmqpExtIntegrationTest.php
@@ -95,7 +95,7 @@ class AmqpExtIntegrationTest extends TestCase
         $envelope = $envelopes[0];
         $newEnvelope = $envelope
             ->with(new DelayStamp(2000))
-            ->with(new RedeliveryStamp(1, 'not_important'));
+            ->with(new RedeliveryStamp(1));
         $sender->send($newEnvelope);
         $receiver->ack($envelope);
 
@@ -174,7 +174,8 @@ class AmqpExtIntegrationTest extends TestCase
 Get envelope with message: Symfony\Component\Messenger\Tests\Fixtures\DummyMessage
 with stamps: [
     "Symfony\\Component\\Messenger\\Transport\\AmqpExt\\AmqpReceivedStamp",
-    "Symfony\\Component\\Messenger\\Stamp\\ReceivedStamp"
+    "Symfony\\Component\\Messenger\\Stamp\\ReceivedStamp",
+    "Symfony\\Component\\Messenger\\Stamp\\ConsumedByWorkerStamp"
 ]
 Done.