+++ src/Symfony/Component/Messenger/Tests/Transport/Sender/SendersLocatorTest.php
@@ -14,7 +14,6 @@ namespace Symfony\Component\Messenger\Tests\Transport\Sender;
 use PHPUnit\Framework\TestCase;
 use Psr\Container\ContainerInterface;
 use Symfony\Component\Messenger\Envelope;
-use Symfony\Component\Messenger\Exception\UnknownSenderException;
 use Symfony\Component\Messenger\Tests\Fixtures\DummyMessage;
 use Symfony\Component\Messenger\Tests\Fixtures\SecondMessage;
 use Symfony\Component\Messenger\Transport\Sender\SenderInterface;
@@ -34,35 +33,6 @@ class SendersLocatorTest extends TestCase
 
         $this->assertSame(['my_sender' => $sender], iterator_to_array($locator->getSenders(new Envelope(new DummyMessage('a')))));
         $this->assertSame([], iterator_to_array($locator->getSenders(new Envelope(new SecondMessage()))));
-    }
-
-    public function testGetSenderByAlias()
-    {
-        $sender1 = $this->getMockBuilder(SenderInterface::class)->getMock();
-        $sender2 = $this->getMockBuilder(SenderInterface::class)->getMock();
-        $sendersLocator = $this->createContainer([
-            'sender1' => $sender1,
-            'sender2' => $sender2,
-        ]);
-
-        $locator = new SendersLocator([], $sendersLocator);
-
-        $this->assertSame($sender1, $locator->getSenderByAlias('sender1'));
-        $this->assertSame($sender2, $locator->getSenderByAlias('sender2'));
-    }
-
-    public function testGetSenderByAliasThrowsException()
-    {
-        $this->expectException(UnknownSenderException::class);
-        $this->expectExceptionMessage('Unknown sender alias');
-
-        $sender1 = $this->getMockBuilder(SenderInterface::class)->getMock();
-        $sendersLocator = $this->createContainer([
-            'sender1' => $sender1,
-        ]);
-
-        $locator = new SendersLocator([], $sendersLocator);
-        $locator->getSenderByAlias('sender2');
     }
 
     private function createContainer(array $senders): ContainerInterface