+++ src/Symfony/Component/Messenger/Transport/Sender/SendersLocator.php
@@ -14,7 +14,6 @@ namespace Symfony\Component\Messenger\Transport\Sender;
 use Psr\Container\ContainerInterface;
 use Symfony\Component\Messenger\Envelope;
 use Symfony\Component\Messenger\Exception\RuntimeException;
-use Symfony\Component\Messenger\Exception\UnknownSenderException;
 use Symfony\Component\Messenger\Handler\HandlersLocator;
 
 /**
@@ -69,14 +68,5 @@ class SendersLocator implements SendersLocatorInterface
                 }
             }
         }
-    }
-
-    public function getSenderByAlias(string $alias): SenderInterface
-    {
-        if ($this->sendersLocator->has($alias)) {
-            return $this->sendersLocator->get($alias);
-        }
-
-        throw new UnknownSenderException(sprintf('Unknown sender alias "%s".', $alias));
     }
 }