+++ src/Symfony/Component/Messenger/CHANGELOG.md
@@ -10,7 +10,6 @@ CHANGELOG
 4.4.0
 -----
 
- * [BC BREAK] The `SyncTransport` and `SyncTransportFactory` classes were removed.
  * Deprecated passing a `ContainerInterface` instance as first argument of the `ConsumeMessagesCommand` constructor,
    pass a `RoutableMessageBus`  instance instead.
  * Added support for auto trimming of Redis streams.
@@ -18,6 +17,12 @@ CHANGELOG
  * Made all dispatched worker event classes final.
  * Added support for `from_transport` attribute on `messenger.message_handler` tag.
  * Added support for passing `dbindex` as a query parameter to the redis transport DSN.
+ * [BC BREAK] Removed `SendersLocatorInterface::getSenderByAlias` added in 4.3.
+ * [BC BREAK] Removed `$retryStrategies` argument from `Worker::__construct`.
+ * [BC BREAK] Removed `$retryStrategyLocator` argument from `ConsumeMessagesCommand::__construct`.
+ * [BC BREAK] Removed `$senderClassOrAlias` argument from `RedeliveryStamp::__construct`.
+ * [BC BREAK] Removed `UnknownSenderException`.
+ * The component is not marked as `@experimental` anymore.
 
 4.3.0
 -----